requirements.txt
# Web Framework
flask
flask-cors
gunicorn

# Google Docs API
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

# Document Processing
beautifulsoup4
lxml

# Vector Store and Embeddings
chromadb
sentence-transformers

# LLM Integration
groq
openai

# Utilities
python-dotenv
requests
tiktoken

# Minimal PyTorch for embeddings (CPU only)
--extra-index-url https://download.pytorch.org/whl/cpu
torch